SSGC WordPress status, in README.md
The images I generated (from the .docx Tom sent) are not great.  I'll try again next week.
Getting the sizes right is a fool's errand, because the ShoreDitch theme seems to dynamically zoom them to some predetermined percentage of the screen.
Have not figured out how to put a featured image on the "Blog Posts" page.